+++ app/src/main/java/com/pepetech/squadhouse/activities/ClubActivity.java
@@ -1,0 +1,20 @@
+package com.pepetech.squadhouse.activities;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.os.Bundle;
+
+import com.pepetech.squadhouse.R;
+
+import com.pepetech.squadhouse.R;
+
+import com.pepetech.squadhouse.R;
+
+public class ClubActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_club);
+    }
+}
