+++ app/src/main/java/com/pepetech/squadhouse/activities/MyProfile/adapters/MyClubsAdapter.java
@@ -14,7 +14,6 @@ import androidx.recyclerview.widget.RecyclerView;
 import com.bumptech.glide.Glide;
 import com.parse.ParseFile;
 import com.pepetech.squadhouse.R;
-import com.pepetech.squadhouse.activities.Club.ClubCreateActivity;
 import com.pepetech.squadhouse.activities.Explore.ExploreClubActivity;
 import com.pepetech.squadhouse.models.Club;
 
@@ -85,9 +84,7 @@ public class MyClubsAdapter extends RecyclerView.Adapter<MyClubsAdapter.ViewHold
                 ivClubImage.setOnClickListener(new View.OnClickListener() {
                     @Override
                     public void onClick(View v) {
-                        // TODO: add routing to create club activity
-                        Intent i = new Intent(v.getContext(), ClubCreateActivity.class);
-                        v.getContext().startActivity(i);
+                        // TODO: add routing to club profile here
 //                        Toast.makeText(context, "Create Club Activity clicked!",
 //                                Toast.LENGTH_SHORT).show();
                     }
