+++ app/src/main/java/com/pepetech/squadhouse/activities/MyProfile/update/UpdateProfileImageActivity.java
@@ -6,6 +6,7 @@ import android.util.Log;
 import android.view.View;
 import android.widget.Button;
 import android.widget.TextView;
+import android.widget.Toast;
 
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.AppCompatImageView;
@@ -14,7 +15,7 @@ import com.bumptech.glide.Glide;
 import com.parse.ParseFile;
 import com.parse.ParseUser;
 import com.pepetech.squadhouse.R;
-import com.pepetech.squadhouse.activities.MyProfile.fragments.UserPhotoUploadBottomSheetDialog;
+import com.pepetech.squadhouse.activities.MyProfile.fragments.PhotoUploadBottomSheetDialog;
 import com.pepetech.squadhouse.models.User;
 
 public class UpdateProfileImageActivity extends AppCompatActivity {
@@ -60,7 +61,7 @@ public class UpdateProfileImageActivity extends AppCompatActivity {
 //                t.show();
                 Log.i(TAG, "Profile Image clicked!");
                 // TODO: fix broken navigation to previous activity
-                UserPhotoUploadBottomSheetDialog bottomSheet = new UserPhotoUploadBottomSheetDialog();
+                PhotoUploadBottomSheetDialog bottomSheet = new PhotoUploadBottomSheetDialog();
                 bottomSheet.show(getSupportFragmentManager(), "ModalBottomSheet");
             }
         });
